Encoding from Binary files cant be retrieved this leads to position of
Parse method throws an NPE when file is no ASCII e.g. jpg. . BTW DXF 
files can be binary e.g. 
http://www.autodesk.com/techpubs/autocad/acad2000/dxf/binary_dxf_files_dxf_aa.htm. 
So right now our feature only supports ASCII dxf files

As this finally block is executed as part of the exception handling for the IOException
thrown in line 128, a ParseException is already generated. Just fixing the NPE by adding
if (in != null)